Change  default settings method to setup and use touchConfig
If a plugin has a method `setup()`, it is called when the plugin is enabled. This is the place where config settings usually are set with sane default values.

The function `touchConfig($configKey, $value)` ensures that a config key is set. If it does not exist yet, it will gets assigned $value, if it already exists, the config value is not changed.
Instead of using one config key after the other, an array can be provided.